feat(llm): refactor clear customized tool
